requirements.txt
langchain
langchain-core
langchain-community
pypdf 
pymupdf
sentence-transformers
faiss-cpu
chromadb
langchain-groq
python-dotenv
pytesseract
Pillow